summary.sim() now produces correct fZd and lHd fields
Previously, the summary function for sim-class objects failed to
produce correct information regarding the first day a second Z was
observed and the last day a H was in the population. This has been
fixed. Additionally, the summary method for sim-class objects now
also computes the days at which the proportion of Zs in the
population exceeds percentiles (in 10% increments).