Populations can be generated with random initial values
Before this fix, all new populations used to be initialized with
only two initial status: 0 (for Hs) or 1 (for Zs). Now, the two new
arguments h.spread and z.spread determine the minimum and maximum
of uniform distributions that are used to generate the Hs' and Zs'
respective status. This is espescially useful for more realistic
(ie. complex) simulations that do not start at a clean state.